requirements.test.txt
aiofiles
pytest
pytest-cov
pytest-homeassistant-custom-component
